feat(play-preview-button.tsx): created PlayPreviewButton
Created PlayPreviewButton to encapsulate a fileRef and PlayButton. Updated FileCard and
FileSelectMenuItem to use the new component.

resolves #174
